test(search): Move tests for OA into a new file.
